Move SidecarApiFactory to separate file
